requirements.txt
numpy 
pandas
matplotlib
scikit-learn
pymongo
pymongo[srv]==3.12
mlflow
joblib 
pyYAML
dotenv
ipykernel
certifi
python-box
ensure
# -e .